ShopifyAPIs/requirements.txt
# pip3 install -r requirements.txt --ignore-installed --no-cache-dir
pip
mysql-connector-python==8.0.23
psycopg2-binary
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.1
defusedxml==0.7.1
et-xmlfile==1.1.0
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
oauthlib==3.2.2
openpyxl==3.0.10
pandas
Pillow
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
twilio==7.14.0
uritemplate==4.1.1
urllib3==1.26.12
XlsxWriter==3.0.6
Flask
Flask-JWT
Flask-JWT-Extended
qrcode
uvicorn
fastapi
asyncio
aio-pika
pydantic
memory_profiler
datetime
psycopg2
boto3
reportlab
python-multipart
